we discuss about class method in python 🐍
A class method in Python is a function that belongs to the class rather than to any specific instance. It can access and modify class-level attributes and variables. Class methods are defined using the @classmethod decorator and take the class itself (usually named cls) as the first argument, allowing them to perform operations that are not tied to any specific instance.